net/rpc/jsonrpc: improve JSON-RPC 1.0 params compatibility
According to JSON-RPC 1.0 specification the parameters property should
be an array of objects. The canonical example of a remote procedure call
uses a simple array of strings:
{"method": "concat", "params": ["foo", "bar"], "id": 1}
Making such a call was not possible with GO's implementation of jsonrpc
as we would always wrap the parameters in an additional one element
array:
{"method": "concat", "params": [["foo", "bar"]], "id": 1}
This works fine as long as both ends use the GO's implementation but
is not compatible with other JSON-RPC 1.0 servers. This commit fixes
this incompatibility.